implemented security patch check
