+++ android/src/main/java/io/parity/signer/screens/initial/eachstartchecks/osversion/ErrorWrongUpdateScreen.kt
@@ -1,10 +1,7 @@
 package io.parity.signer.screens.initial.eachstartchecks.osversion
 
 import android.content.res.Configuration
-import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -13,48 +10,30 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.SecurityUpdate
-import androidx.compose.material.icons.outlined.SettingsBackupRestore
-import androidx.compose.material.icons.outlined.UpdateDisabled
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.ExperimentalTextApi
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.withAnnotation
-import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import io.parity.signer.R
 import io.parity.signer.components.base.PrimaryButtonWide
 import io.parity.signer.domain.Callback
-import io.parity.signer.screens.scan.errors.COMPOSE_URL_TAG_ANNOTATION
 import io.parity.signer.ui.theme.SignerNewTheme
 import io.parity.signer.ui.theme.SignerTypeface
-import io.parity.signer.ui.theme.fill12
-import io.parity.signer.ui.theme.fill6
-import io.parity.signer.ui.theme.pink300
 import io.parity.signer.ui.theme.pink500
-import io.parity.signer.ui.theme.textSecondary
 import io.parity.signer.ui.theme.textTertiary
 
 @Composable
 fun WrongOsVersionNotificationScreen(
-	currentOSVersion: String,
-	minimalOsVersion: String,
+	exposedVulnarabilities: List<KnownOSIssue>,
 	onProceed: Callback
 ) {
 	Column(
@@ -90,8 +69,7 @@ fun WrongOsVersionNotificationScreen(
 					.fillMaxWidth(1f),
 				text = stringResource(
 					R.string.initial_screen_outdated_os_description,
-					currentOSVersion,
-					minimalOsVersion,
+					exposedVulnarabilities.joinToString(", ") { it.toString() }
 				),
 				color = MaterialTheme.colors.textTertiary,
 				style = SignerTypeface.BodyL,
@@ -124,7 +102,7 @@ private fun WrongOsVersionNotificationScreenPreview() {
 	Box(modifier = Modifier.fillMaxSize()) {
 		SignerNewTheme() {
 			WrongOsVersionNotificationScreen(
-				"6.0", "7.1",
+				listOf(KnownOSIssue.CVE_2022_20465),
 				onProceed = {},
 			)
 		}
