+++ android/src/main/res/values/strings.xml
@@ -455,7 +455,7 @@
 	<string name="add_derived_keys_screen_alert_title">Confirm by entering the PIN code</string>
 	<string name="add_derived_keys_screen_alert_content">To finalize the addition of keys to the key set, please confirm that you have scanned the QR code back into the app by entering the PIN code.</string>
 	<string name="initial_screen_outdated_os_title">Consider updating operation system before using vault</string>
-	<string name="initial_screen_outdated_os_description">Current Oos version is %1$s and minimal recommended is %1$s due to known vulnerabilities of old OS versions.</string>
+	<string name="initial_screen_outdated_os_description">Current Android version missing important security updated, exposed to at least: %1$s. It is recommended to install vault on more up to date android version.</string>
 
 </resources>
